+++ packages/rangelink-vscode-extension/src/constants/index.ts
@@ -3,3 +3,4 @@ export * from './commandIds';
 export * from './contextKeys';
 export * from './settingDefaults';
 export * from './settingKeys';
+export * from './vscodeCommandIds';